Move pytest fixtures to main_node conftest.py
