+++ main_node/transfer_learning/tests/test_mtl.py
@@ -1,7 +1,5 @@
 from typing import Dict, Tuple
 from copy import deepcopy
-import pytest
-from unittest.mock import MagicMock
 
 from core_entities.configuration import Configuration
 from core_entities.experiment import Experiment
@@ -14,13 +12,6 @@ from tools.restore_db import RestoreDB
 
 experiment_description_file = "./Resources/tests/test_cases_product_configurations/test_case_0.json"
 rdb = RestoreDB()
-
-@pytest.fixture(autouse=True)
-def mock_configuration_selection_event_service(monkeypatch):
-    mock_connection_thread = MagicMock()
-    monkeypatch.setattr(ConfigurationSelection, '_EventServiceConnection', MagicMock(return_value=mock_connection_thread))
-    monkeypatch.setattr('configuration_selection.configuration_selection.publish', MagicMock())
-    return mock_connection_thread
 
 class TestMTL:
     def test_0(self, get_workers, get_configurations_2_float):
